+++ CHANGELOG.md
@@ -1,5 +1,11 @@
 # Changelog
 
+# v24.32.0
+
+- Add dependabot config.
+- Add debug logging to `get_latest_log_file` function
+- Fix test for `get_latest_log_file` function to clear the log directory before running the test
+
 # v24.30.0
 
 - Add `supportsPosixRename` option to SFTP source so that post copy actions will work on server that don't support POSIX renames